Working on the design of the project system.
Got the different specification files reviewed by Peter (Specification.docx, User specification.docx, Kernel specification.docx and Project specification.docx). All changes made by Peter have been merged back. Based on the specifications started working on the UML designs. The design for the scheduling and the variables is very close to done. The design of the dataset and the dataset application is about 50% done. The other designs are improving but still a bit from being done.