IssueId #51: Unify the csproj files
Updated all the csproj files to link to a single baseconfiguration.target file for all the base configurations. The BaseConfiguration.targets file contains:
- The default settings for all the configurations. Being:
  - ProductVersion
  - SchemaVersion
  - Configuration: defaults to debug
  - TargetFrameworkVersion: defaults to 4.0
  - FileAlignment
  - AppDesignerFolder: Defaults to 'Properties'
  - OutputPath: Defaults to build\bin\<PLATFORM>\<CONFIGURATION>
  - ErrorReport
  - WarningLevel: defaults to 4
  - TreatWarningsAsErrors: Defaults to true
- The standard configurations for the 'Any CPU' platform and the 'Debug'and 'Release' configurations.

The individual project files need to minimally define:
- SolutionDir
- ProjectGuid
- OutputType
- RootNamespace
- AssemblyName

Assemblies can chose to not be signed by setting the 'ShouldSign' property to false. And XML documentation can be turned off by setting the 'ShouldDoDocumentation' property to false