[ap-66] Synchronize history roll-back and roll-forward with the online state of the datasets
Implemented a mechanism that allows a dataset to be unloaded / loaded in reaction to updates to the timeline. Note that at the moment we don't block roll-back / roll-forward if we load / unload a dataset but maybe we should (because unloading a dataset destroys the undo stack of the dataset itself, which will be related to the undo stack of the project?).